+++ src/documentation_glossary.jl
@@ -90,6 +90,7 @@ define!(
 define!(:LaTeX, :proj, raw"\operatorname{proj}")
 define!(:LaTeX, :prox, raw"\operatorname{prox}")
 define!(:LaTeX, :quad, raw"\quad")
+define!(:LaTeX, :qquad, raw"\qquad")
 define!(:LaTeX, :reflect, raw"\operatorname{refl}")
 define!(:LaTeX, :retr, raw"\operatorname{retr}")
 define!(:LaTeX, :rm, (letter) -> raw"\mathrm{" * "$letter" * "}")
@@ -198,11 +199,11 @@ define!(
     :Note,
     :ManifoldDefaultFactory,
     (type::String) -> """
-!!! info
-    This function generates a [`ManifoldDefaultsFactory`](@ref) for [`$(type)`](@ref).
-    For default values, that depend on the manifold, this factory postpones the construction
-    until the manifold from for example a corresponding [`AbstractManoptSolverState`](@ref) is available.
-""",
+    !!! info
+        This function generates a [`ManifoldDefaultsFactory`](@ref) for [`$(type)`](@ref).
+        For default values, that depend on the manifold, this factory postpones the construction
+        until the manifold from for example a corresponding [`AbstractManoptSolverState`](@ref) is available.
+    """,
 )
 define!(
     :Note,
@@ -221,11 +222,11 @@ define!(
     :Note,
     :OutputSection,
     (; p_min="p^*") -> """
-# Output
+                                                                     # Output
 
-The obtained approximate minimizer ``$(p_min)``.
-To obtain the whole final state of the solver, see [`get_solver_return`](@ref) for details, especially the `return_state=` keyword.
-""",
+                                                                     The obtained approximate minimizer ``$(p_min)``.
+                                                                     To obtain the whole final state of the solver, see [`get_solver_return`](@ref) for details, especially the `return_state=` keyword.
+                                                                     """,
 )
 define!(
     :Note,
@@ -249,14 +250,14 @@ define!(
     :Problem,
     :Constrained,
     (; M="M", p="p") -> """
-    ```math
-\\begin{aligned}
-$(_tex(:argmin))_{$p ∈ $(_math(:M; M=M))} & f($p)\\\\
-$(_tex(:text, "subject to"))$(_tex(:quad))&g_i($p) ≤ 0 \\quad $(_tex(:text, " for ")) i= 1, …, m,\\\\
-\\quad & h_j($p)=0 \\quad $(_tex(:text, " for ")) j=1,…,n,
-\\end{aligned}
-```
-""",
+                                                                            ```math
+                                                                        \\begin{aligned}
+                                                                        $(_tex(:argmin))_{$p ∈ $(_math(:M; M=M))} & f($p)\\\\
+                                                                        $(_tex(:text, "subject to"))$(_tex(:quad))&g_i($p) ≤ 0 \\quad $(_tex(:text, " for ")) i= 1, …, m,\\\\
+                                                                        \\quad & h_j($p)=0 \\quad $(_tex(:text, " for ")) j=1,…,n,
+                                                                        \\end{aligned}
+                                                                        ```
+                                                                        """,
 )
 define!(
     :Problem,
@@ -279,13 +280,13 @@ define!(
     :Problem,
     :NonLinearLeastSquares,
     (; M="M", p="p") -> """
-```math
-$(_tex(:argmin))_{$p ∈ $(_math(:M; M=M))} $(_tex(:frac,1,2)) $(_tex(:sum))_{i=1}^m $(_tex(:abs, "f_i($p)"))^2
-```
+                                                                        ```math
+                                                                        $(_tex(:argmin))_{$p ∈ $(_math(:M; M=M))} $(_tex(:frac,1,2)) $(_tex(:sum))_{i=1}^m $(_tex(:abs, "f_i($p)"))^2
+                                                                        ```
 
-where ``f: $(_math(:M; M=M)) → ℝ^m`` is written with component functions ``f_i: $(_math(:M; M=M)) → ℝ``, ``i=1,…,m``,
-and each component function is continuously differentiable.
-""",
+                                                                        where ``f: $(_math(:M; M=M)) → ℝ^m`` is written with component functions ``f_i: $(_math(:M; M=M)) → ℝ``, ``i=1,…,m``,
+                                                                        and each component function is continuously differentiable.
+                                                                        """,
 )
 
 #
@@ -438,8 +439,8 @@ define!(
 )
 define!(:Variable, :p, :type, "P")
 define!(:Variable, :p, :default, (; M="M") -> _link(:rand; M=M))
-define!(:Variable, :p, :as_Iterate, "storing the current iterate")
-define!(:Variable, :p, :as_Initial, "to specify the initial value")
+define!(:Variable, :p, :as_Iterate, " storing the current iterate")
+define!(:Variable, :p, :as_Initial, " to specify the initial value")
 
 define!(
     :Variable,
@@ -480,15 +481,6 @@ define!(:Variable, :stopping_criterion, :type, "StoppingCriterion")
 
 define!(
     :Variable,
-    :sub_problem,
-    :description,
-    (; M="M") ->
-        "specify a problem for a solver or a closed form solution function, which can be allocating or in-place.",
-)
-define!(:Variable, :sub_problem, :type, "Union{AbstractManoptProblem, F}")
-
-define!(
-    :Variable,
     :sub_kwargs,
     :description,
     "a named tuple of keyword arguments that are passed to [`decorate_objective!`](@ref) of the sub solvers objective, the [`decorate_state!`](@ref) of the subsovlers state, and the sub state constructor itself.",
@@ -519,10 +511,10 @@ define!(
     :subgrad_f,
     :description,
     (; M="M", p="p") -> """
-the subgradient ``∂f: $(_math(:M; M=M)) → $(_math(:TM; M=M))`` of f as a function `(M, p) -> X`
-or a function `(M, X, p) -> X` computing `X` in-place.
-This function should always only return one element from the subgradient.
-""",
+                                                                        the subgradient ``∂f: $(_math(:M; M=M)) → $(_math(:TM; M=M))`` of f as a function `(M, p) -> X`
+                                                                        or a function `(M, X, p) -> X` computing `X` in-place.
+                                                                        This function should always only return one element from the subgradient.
+                                                                        """,
 )
 define!(
     :Variable,