+++ src/plans/objective.jl
@@ -174,6 +174,7 @@ This function should dispatch on `Val(element)`.
 Currently supported
 * `:Cost` passes to the [`get_cost_function`](@ref)
 * `:Gradient` passes to the [`get_gradient_function`](@ref)
+* `:SubGradient` passes to the [`get_subgradient_function`](@ref)
 """
 set_parameter!(amo::AbstractManifoldObjective, e::Symbol, args...)
 
@@ -183,6 +184,10 @@ function set_parameter!(amo::AbstractManifoldObjective, ::Val{:Cost}, args...)
 end
 function set_parameter!(amo::AbstractManifoldObjective, ::Val{:Gradient}, args...)
     set_parameter!(get_gradient_function(amo), args...)
+    return amo
+end
+function set_parameter!(amo::AbstractManifoldObjective, ::Val{:SubGradient}, args...)
+    set_parameter!(get_subgradient_function(amo), args...)
     return amo
 end