+++ src/plans/plan.jl
@@ -116,6 +116,7 @@ include("cost_plan.jl")
 include("gradient_plan.jl")
 include("hessian_plan.jl")
 include("proximal_plan.jl")
+include("proximal_gradient_plan.jl")
 include("subgradient_plan.jl")
 include("vectorial_plan.jl")